Swap tslint for eslint
1. I can't get the tests to pass
2. tslint doesn't pass
3. tslint is deprecated

I just need SOMETHING to pass so that we have a baseline to run
CI against and then we can work on coverage. I think the fastest
way out is to rip the bandaid and at least get `ng lint` running
in the frontend project.

This is a mix of automated and manual fixes that I felt introduced
the least oppinion and got us to a clean `ng lint` and `ng serve`.