More complete build script
We're fast approaching the need for a "build script" or pipeline.
This adds a cleanup step before the build, then builds both the
app and lib (both needed for the node app to run) and finishes
by moving the lib artifacts to the build dir.

Previously this was being done manually in CI (which is not source
controlled because digital ocean).