Link previews for individual blueprints
Previously we were depending on the fallthrough route to handle
blueprint links. This now loads them from the backend. For humans
we don't do anything new (yet) but for robots we customize the
page meta so that we can get shareable link unfurls for each
blueprint, with it's own thumbnail as preview image.