Replace frontend NG_APP_HOST
To avoid multiple builds (now docker images) for each environment
we will need a way to pass ENV from the backend to the frontend
on page load. But all the internet answers seem awful.
I'd like to delay making a choice so I'm going to replace the
only really important one with runtime code.