Allow TSC to compile js files
We have a smattering of JS files in the backend. Letting TSC see
them has the advanatge of including them in the normal dev cycle
but the disadvantage of exposing us to the amazing slew of
errors from pixi.js

TSC naturally doesn't "excludes" the node_modules direstory, so
all we need to add is /build which is full of our output js

Then we use skipLibCheck to avoid type checking node_modules.