Add extract-export script and seed task
This is a master script to run all the other batch scripts.
Right now it "works" although it doesn't seem to terminate on
it's own (possibly a dangling handler in the icon generation)
and it takes FOREVER, so I'm not going to recommend it as the
default way to seed the project until I can refine the process.