chore: reduce console spam
I'm not sure exactly where all the formatting changes here came from.
I'll be adding some kind of linting soon so that we can avoid this
in the future. For now I'm just gonna let it slide, because it was
likely my local config and that is generally pretty unopinionated.
If it was the AI then I appologise for the churn.